+++ src/main/java/kr/codesquad/controller/LottoController.java
@@ -1,9 +1,9 @@
 package kr.codesquad.controller;
 
-import kr.codesquad.domain.Lotto;
-import kr.codesquad.domain.Lottos;
+import kr.codesquad.domain.lotto.Lotto;
+import kr.codesquad.domain.lotto.Lottos;
 import kr.codesquad.domain.Result;
-import kr.codesquad.domain.WinLotto;
+import kr.codesquad.domain.lotto.WinLotto;
 import kr.codesquad.view.InputView;
 import kr.codesquad.view.OutputView;
