+++ src/main/java/kr/codesquad/domain/lotto/WinLotto.java
@@ -1,4 +1,4 @@
-package kr.codesquad.domain;
+package kr.codesquad.domain.lotto;
 
 public class WinLotto {
