+++ src/main/java/kr/codesquad/domain/lotto/factory/LottoFactory.java
@@ -1,4 +1,4 @@
-package kr.codesquad.factory;
+package kr.codesquad.domain.lotto.factory;
 
 import java.util.List;
